add placeholders for cards + organisms and move existing components to appriopriate directory
