+++ src/components/molecules/SpotlightCard.tsx
