+++ src/components/organisms/faq/component.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const Faq: React.FC = () => (
+  <div>
+    {/* Placeholder for Faq component */}
+  </div>
+);
