+++ src/components/organisms/faq/index.ts
@@ -1,0 +1,2 @@
+export { Faq as default } from './component';
+export * from './component';
