+++ src/components/organisms/footer/component.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const Footer: React.FC = () => (
+  <div>
+    {/* Placeholder for Footer component */}
+  </div>
+);
