+++ src/components/organisms/footer/index.ts
@@ -1,0 +1,2 @@
+export { Footer as default } from './component';
+export * from './component';
