+++ src/components/organisms/navbar/component.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const NavBar: React.FC = () => (
+  <div>
+    {/* NavBar for Footer component */}
+  </div>
+);
