+++ src/components/organisms/navbar/index.ts
@@ -1,0 +1,2 @@
+export { NavBar as default } from './component';
+export * from './component';
